# omborai/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OmborAI API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/stores": {
      "get": {
        "tags": [
          "stores"
        ],
        "summary": "List Stores",
        "operationId": "list_stores_v1_stores_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StoreOut"
                  },
                  "type": "array",
                  "title": "Response List Stores V1 Stores Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "stores"
        ],
        "summary": "Create Store",
        "operationId": "create_store_v1_stores_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/shifts": {
      "post": {
        "tags": [
          "shifts"
        ],
        "summary": "Open Shift",
        "operationId": "open_shift_v1_shifts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftOpenIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/shifts/current": {
      "get": {
        "tags": [
          "shifts"
        ],
        "summary": "Current Shift",
        "operationId": "current_shift_v1_shifts_current_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Store Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/shifts/{shift_id}/close": {
      "post": {
        "tags": [
          "shifts"
        ],
        "summary": "Close Shift",
        "operationId": "close_shift_v1_shifts__shift_id__close_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shift Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftCloseIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftSummaryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ShiftCloseIn": {
        "properties": {
          "closing_cash": {
            "type": "integer",
            "maximum": 1000000000000.0,
            "minimum": 0.0,
            "title": "Closing Cash"
          }
        },
        "type": "object",
        "required": [
          "closing_cash"
        ],
        "title": "ShiftCloseIn"
      },
      "ShiftOpenIn": {
        "properties": {
          "store_id": {
            "type": "string",
            "format": "uuid",
            "title": "Store Id"
          },
          "opening_cash": {
            "type": "integer",
            "maximum": 1000000000000.0,
            "minimum": 0.0,
            "title": "Opening Cash",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "store_id"
        ],
        "title": "ShiftOpenIn"
      },
      "ShiftOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "store_id": {
            "type": "string",
            "format": "uuid",
            "title": "Store Id"
          },
          "opened_by": {
            "type": "string",
            "format": "uuid",
            "title": "Opened By"
          },
          "opened_at": {
            "type": "string",
            "format": "date-time",
            "title": "Opened At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "opening_cash": {
            "type": "integer",
            "title": "Opening Cash"
          },
          "closing_cash": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closing Cash"
          }
        },
        "type": "object",
        "required": [
          "id",
          "store_id",
          "opened_by",
          "opened_at",
          "closed_at",
          "opening_cash",
          "closing_cash"
        ],
        "title": "ShiftOut"
      },
      "ShiftSummaryOut": {
        "properties": {
          "shift": {
            "$ref": "#/components/schemas/ShiftOut"
          },
          "sales_count": {
            "type": "integer",
            "title": "Sales Count"
          },
          "total_sales": {
            "type": "integer",
            "title": "Total Sales"
          },
          "refunds_count": {
            "type": "integer",
            "title": "Refunds Count"
          },
          "total_refunds": {
            "type": "integer",
            "title": "Total Refunds"
          },
          "by_method": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Method"
          },
          "expected_cash": {
            "type": "integer",
            "title": "Expected Cash"
          },
          "closing_cash": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closing Cash"
          },
          "difference": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difference"
          }
        },
        "type": "object",
        "required": [
          "shift",
          "sales_count",
          "total_sales",
          "refunds_count",
          "total_refunds",
          "by_method",
          "expected_cash",
          "closing_cash",
          "difference"
        ],
        "title": "ShiftSummaryOut"
      },
      "StoreIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "StoreIn"
      },
      "StoreOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "StoreOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}